Rename 'fixed_rate_shards' dir to 'parda_shards'
