+++ bench/mrc_test/mrc_buckets_test.c
@@ -6,7 +6,7 @@
 #include "random/zipfian_random.h"
 #include "test/mytester.h"
 
-#include "fixed_rate_shards/parda_fixed_rate_shards.h"
+#include "parda_shards/parda_fixed_rate_shards.h"
 #include "shards/fixed_size_shards.h"
 #include "mimir/buckets.h"
 #include "mimir/mimir.h"
